src/requirements.txt
# ─── Core ───
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
pandas>=2.0.0,<3.0.0

# ─── LangChain Ecosystem ───
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.3.0
langchain-google-genai>=2.0.0
langchain-google-community[calendar]>=3.0.0
langchainhub
langsmith

# ─── LangGraph ───
langgraph>=0.2.0

# ─── Deep Agents ───
deepagents>=0.4.0

# ─── RAG / Vector DB ───
chromadb>=0.5.0
langchain-chroma
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
pypdf>=4.0.0

# ─── Frontend ───
streamlit==1.54.0

# ─── APIs ───
openai>=1.0.0
tavily-python>=0.7.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# ─── Optimization ───
ortools>=9.10.0